+++ CHANGELOG.md
@@ -19,13 +19,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Custom onLoadUrlInBrowser and onLoadUrlInBrowserError props
 
+### Changed
+
+- Now opening oauth tabs inside the app instead of in a separate browser. This simplifies the configuration process and makes it easier for the user to get back to the app in case of failures.
+
 ## [3.0.0]
 
 ### Removed
 
 - No longer rendering the widgets inside of a SafeAreaView. Users of the widgets need to handle this.
 
-### Updated
+### Changed
 
 - Stopped using the deprecated react-native SafeAreaView in favor of react-native-safe-area-context